Expand the set of characters allowed in words
Before this commit, words could only contain alphabetic and numeric characters.
This is unduly limiting; for example, combining marks (like accents) are not included.
This commit expands the set of allowed characters based on the recommendations of
UAX 31 <https://www.unicode.org/reports/tr31/>, with a bias toward allowing more characters
(though emoji are excluded).
In addition, unassigned and private use characters are assumed to be allowed in words
(ensuring that case conversion will pass them through unchanged).

This change means we can no longer rely only on the Unicode data tables
shipped with the standard library.
A new `tables` binary crate is in charge of generating the tables we need
(which consumes 3600 bytes of data).